pipelines/OCR_gemini/requirements.txt
google-cloud-aiplatform>=1.38.0
vertexai>=0.0.1
python-dotenv>=1.0.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
pandas>=2.0.0
